more init into config
