+++ src/VipsObject.php
@@ -68,10 +68,10 @@ abstract class VipsObject extends GObject
 
     public function __construct($pointer)
     {
-        $this->pointer = Init::ffi()->
-            cast(Init::ctypes("VipsObject"), $pointer);
-        $this->gObject = Init::ffi()->
-            cast(Init::ctypes("GObject"), $pointer);
+        $this->pointer = Config::ffi()->
+            cast(Config::ctypes("VipsObject"), $pointer);
+        $this->gObject = Config::ffi()->
+            cast(Config::ctypes("GObject"), $pointer);
 
         parent::__construct($pointer);
     }
@@ -79,12 +79,12 @@ abstract class VipsObject extends GObject
     // print a table of all active vipsobjects ... handy for debugging
     public static function printAll()
     {
-        Init::ffi()->vips_object_print_all();
+        Config::ffi()->vips_object_print_all();
     }
 
     public function getDescription()
     {
-        return Init::ffi()->vips_object_get_description($this->pointer);
+        return Config::ffi()->vips_object_get_description($this->pointer);
     }
 
     // get the pspec for a property
@@ -93,10 +93,10 @@ abstract class VipsObject extends GObject
     // FIXME add a cache for this thing
     public function getPspec(string $name)
     {
-        $pspec = Init::ffi()->new("GParamSpec*[1]");
-        $argument_class = Init::ffi()->new("VipsArgumentClass*[1]");
-        $argument_instance = Init::ffi()->new("VipsArgumentInstance*[1]");
-        $result = Init::ffi()->vips_object_get_argument(
+        $pspec = Config::ffi()->new("GParamSpec*[1]");
+        $argument_class = Config::ffi()->new("VipsArgumentClass*[1]");
+        $argument_instance = Config::ffi()->new("VipsArgumentInstance*[1]");
+        $result = Config::ffi()->vips_object_get_argument(
             $this->pointer,
             $name,
             $pspec,
@@ -118,7 +118,7 @@ abstract class VipsObject extends GObject
         $pspec = $this->getPspec($name);
         if (\FFI::isNull($pspec)) {
             # need to clear any error, this is horrible
-            Init::ffi()->vips_error_clear();
+            Config::ffi()->vips_error_clear();
             return 0;
         } else {
             return $pspec->value_type;
@@ -128,13 +128,13 @@ abstract class VipsObject extends GObject
     public function getBlurb(string $name): string
     {
         $pspec = $this->getPspec($name);
-        return Init::ffi()->g_param_spec_get_blurb($pspec);
+        return Config::ffi()->g_param_spec_get_blurb($pspec);
     }
 
     public function getArgumentDescription(string $name): string
     {
         $pspec = $this->getPspec($name);
-        return Init::ffi()->g_param_spec_get_description($pspec);
+        return Config::ffi()->g_param_spec_get_description($pspec);
     }
 
     public function get(string $name)
@@ -142,7 +142,7 @@ abstract class VipsObject extends GObject
         $gvalue = new GValue();
         $gvalue->setType($this->getType($name));
 
-        Init::ffi()->
+        Config::ffi()->
             g_object_get_property($this->gObject, $name, $gvalue->pointer);
         $value = $gvalue->get();
 
@@ -159,13 +159,13 @@ abstract class VipsObject extends GObject
         $gvalue->setType($this->getType($name));
         $gvalue->set($value);
 
-        Init::ffi()->
+        Config::ffi()->
             g_object_set_property($this->gObject, $name, $gvalue->pointer);
     }
 
     public function setString(string $string_options)
     {
-        $result = Init::ffi()->
+        $result = Config::ffi()->
             vips_object_set_from_string($this->pointer, $string_options);
 
         return $result == 0;
@@ -173,7 +173,7 @@ abstract class VipsObject extends GObject
 
     public function unrefOutputs()
     {
-        Init::ffi()->vips_object_unref_outputs($this->pointer);
+        Config::ffi()->vips_object_unref_outputs($this->pointer);
     }
 }
